requirements.txt
solana==0.8.0
python-dotenv==0.17.1
pydash
